--- lib/application/theme/theme.dart
@@ -1,2 +1,0 @@
-export 'enums/app_theme.dart';
-export 'cubit/theme_cubit.dart';
