+++ lib/presentation/app_widget.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../application/theme/theme.dart';
 import '../apps_list/apps_list.dart';
+import '../theme/theme.dart';
 import 'loading_page.dart';
 import '../logs/log_page.dart';
 import '../settings/settings_page.dart';
