+++ lib/settings/cubit/settings_cubit.dart
@@ -7,7 +7,7 @@ import 'package:flutter/material.dart';
 import 'package:nyrna/application/helpers/json_converters.dart';
 import 'package:nyrna/infrastructure/icon_manager/icon_manager.dart';
 import 'package:nyrna/settings/settings_service.dart';
-import 'package:nyrna/presentation/styles.dart';
+import 'package:nyrna/theme/styles.dart';
 import 'package:window_size/window_size.dart' as window;
 
 import '../../../infrastructure/launcher/src/hotkey.dart';
