+++ lib/settings/widgets/behaviour_section.dart
@@ -2,7 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
 import '../../presentation/core/core.dart';
-import '../../presentation/styles.dart';
+import '../../theme/styles.dart';
 import '../cubit/settings_cubit.dart';
 
 class BehaviourSection extends StatelessWidget {
