+++ lib/settings/widgets/theme_section.dart
@@ -4,7 +4,7 @@ import 'package:flex_color_picker/flex_color_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../application/theme/theme.dart';
+import '../../theme/theme.dart';
 import '../settings.dart';
 
 class ThemeSection extends StatelessWidget {
