+++ lib/theme/cubit/theme_cubit.dart
@@ -1,8 +1,9 @@
 import 'package:bloc/bloc.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter/material.dart';
-import 'package:nyrna/application/theme/enums/app_theme.dart';
 import 'package:nyrna/settings/settings_service.dart';
+
+import '../theme.dart';
 
 part 'theme_state.dart';
